# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Photo Album API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '402':
          description: The upload would exceed the organization's storage limit (storage_limit_exceeded).
        '413':
          description: The declared size exceeds the upload limit (file_too_large).
        '500':
          description: Failed to generate upload URL.
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List asset library items (brand assets and reference works)
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [brand, reference]
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
        "401":
          description: Unauthorized
    post:
      operationId: createAsset
      tags: [assets]
      summary: Register an uploaded file as an asset (upload the file via the storage request-url flow first)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /assets/{id}:
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset's metadata (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    Asset:
      type: object
      required: [id, kind, name, storageKey, contentType, createdAt, updatedAt]
      properties:
        id:
          type: integer
        kind:
          type: string
          enum: [brand, reference]
        name:
          type: string
        isPrimary:
          type: boolean
          description: The designated primary logo for its scope (the organization, or its project) — preferred by Create and the only logo Campaigns attach automatically (#206).
        variant:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        folder:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        storageKey:
          type: string
        contentType:
          type: string
        filename:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        createdById:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AssetInput:
      type: object
      required: [kind, name, storageKey, contentType]
      properties:
        kind:
          type: string
          enum: [brand, reference]
        name:
          type: string
          minLength: 1
        variant:
          type: string
        notes:
          type: string
        folder:
          type: string
        projectId:
          type: integer
        storageKey:
          type: string
          description: The objectPath returned by the storage request-url flow (must start with /objects/)
        contentType:
          type: string
        filename:
          type: string
        fileSize:
          type: integer
    AssetUpdate:
      type: object
      properties:
        kind:
          type: string
          enum: [brand, reference]
        name:
          type: string
          minLength: 1
        variant:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        folder:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        isPrimary:
          type: boolean
          description: Mark (true) or unmark (false) as the primary logo for the asset's scope; replaces any previous primary in that scope. Organization owners/admins only; brand image assets only.
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath, uploadHeaders]
      properties:
        uploadURL:
          type: string
          description: Presigned URL for PUT upload. Absolute in production; may be a relative proxy path (e.g. `/gcs/...`) in local dev.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        uploadHeaders:
          type: object
          additionalProperties:
            type: string
          description: Headers the PUT must send exactly — they're signed into the URL (Content-Type, and x-goog-content-length-range when size signing is enabled).
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...