# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Photo Album API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from "./uploadUrlRequest";
import type { UploadUrlResponseUploadHeaders } from "./uploadUrlResponseUploadHeaders";

export interface UploadUrlResponse {
  /** Presigned URL for PUT upload. Absolute in production; may be a relative proxy path (e.g. `/gcs/...`) in local dev. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  /** Headers the PUT must send exactly — they're signed into the URL (Content-Type, and x-goog-content-length-range when size signing is enabled). */
  uploadHeaders: UploadUrlResponseUploadHeaders;
  metadata?: UploadUrlRequest;
}